hpcrun: Remove LUSH agents and unwinding
The LUSH subsystem is a model and interface for measuring "logical"
data, while it has some great ideas it has been pushed aside in various
refactors and new (entirely different) "logical" measurement subsystems.
At this point it's bitrotting in the corner, it's time to sweep it up.

This commit removes most of the LUSH pieces in hpcrun, particularly the
agents and unwinding fragments. These pieces are dead code, they can't
be activated anymore or if they were everything would crash. This makes
them save to remove without causing backwards compatibility problems.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>