hpcrun: Improve handling when CUDA is unavailable
If the CUDA driver (libcuda.so) is unavailable, currently `hpcrun` will
error with an `abort()` (previously `exit(-1)`). It turns out that the
first time the missing driver produces an error is when the foil is
loaded into the process address space.

Adjust the CUDA/CUPTI foil to return semi-reasonable return codes when
the foil fails to load for any reason.

Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/842

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>