devcontainers: Remove predeps-based devcontainers
Most of the devcontainers are based on the ci.predeps containers, that
is they reflect the environments tested in CI/CD. There are a lot of
them, and there is no good way to keep them in-sync with the actual CI
pipeline where they are used. It also means these containers do
"double-duty" for remarkably conflicting use cases.

This commit removes these devcontainers. The only devcontainer now is
the "ubercontainer" which is most suited for development.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>