Update MPI support for Meson 1.6.0
Meson 1.6.0 includes a serious update to the `dependency('MPI')` logic
that adds MPICH support and changes the search precedence. Unfortunately
it also includes a bug in said logic that breaks our build when MPICH
is being used that we don't detect.

This commit updates the MPI dependency logic to exploit the enhanced
upstream support and work around this bug. The fallback logic used to
work around missing support pre-1.6.0 is now deprecated and only used
when running an older version of Meson.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>