tests: Remove hpcprof-mpi tests
The test suite (`meson test`) has long included tests for `hpcprof-mpi`,
the MPI-enabled variant of `hpcprof`. The MPI implementations are unkind
to tests, in CI they crash and hang due to no fault of our own (AFAICT).
Even when they work, they are serialized (to avoid MPI bugs) and
significantly slow down the testing process on multi-core machines.
It's time to put a complete stop to the madness.

This commit removes the `hpcprof-mpi` tests and the exceptions collected
in the CI to dodge all the bugs they have.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>