Force version of extra packages in Intel predeps
The Intel repository now ships bits from OneAPI 2025, a major version
bump that is incompatible with the older OneAPI 2024 we test against.
Fix the versions of affected "extra" packages to avoid APT version
resolution conflicts.

The OneAPI version is now hardcoded in the Containerfile to avoid
unexpected errors if the version argument is bumped.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>