Move vendored Python dependencies to subproject
The REUSE standard implicitly considers subprojects (`subprojects/*/`)
as separate "projects" from the main repository. This makes them
particularly useful for expressing external sources that we vendor into
our repository for convenience's sake.

Move the vendored Python wheels and related code to
`subprojects/hpctesttool-wheels/`, taking advantage of this feature to
clean up a number of LICENSES in the top project.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>