Remove the tbb-wrap subproject
The tbb-wrap/ subproject is a Meson subproject only because of how
dependency fallbacks work by default, it doesn't actually represent a
separate "Project" as defined by REUSE. To avoid confusion remove this
"subproject" and restore logic in the top-level meson.build.

Note that the build will now fail if the TBB wrap is used on an older
version of Meson (before 1.6.x). This is within reason, wraps are
designed for the latest Meson without so much attention to
backwards-compatibility. A Meson warning() is emitted when this scenario
is detected to ensure users are aware of the issue.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>