ci: Split script into multiple fragments
The .gitlab-ci.yml script is much too large, there's a lot of complex
logic and multiple pipelines that span the ~1K LoC of this file. This
commit splits this up into multiple fragments ci/*.gitlab-ci.yml, which
are then composed by the top-level .gitlab-ci.yml script.

Note that _most_ rules based on whether the pipeline is running in a
fork or not are now handled during the composition rather than in the
jobs themselves. This helps remove some complexity in already complex
pipelines.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>