ci: Combine based on pipeline location (site)
The main limiting factor for a fragment is what CI runners are
available,
i.e. what tags (or none) we can put on jobs and expect to get a working
pipeline. Instead of guesstimating based on the MR status, this commit
replaces that logic with new logic based entirely on the FQDN of the
GitLab instance.

Currently a CI pipeline run on an unknown GitLab instance is empty, i.e.
has no jobs. This should be refined later to include some set of fast
and extremely common checks, that make as few assumptions about the CI
runners as possible.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>