ci: Add site pipeline for HPSF CI
The HPSF has granted us access to their CI resources, so we should use
them to their fullest to improve our CI pipelines. This commit adds a
new fragment that runs all the jobs we currently need hardware at Rice
to run, such as AMD or Intel GPUs.

Unlike the main GitLab pipeline(s), HSPF CI doesn't use ci.predeps to
prep the environment, instead we use scripts that run `apt-get install`
during the CI job itself. This is necessary because the HPSF GitLab
instance doesn't include the container registry, so we don't have a
place to store images constructed on-the-fly.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>