ci/hpsf: Also test on PowerPC
The HPSF CI now exposes their PowerPC resources for our use. This commit
adds the CI jobs to test on those resources, using the same environments
and settings as what we test on the Rice-hosted PowerPC runners.

Note that the same Python and PAPI bugs are still present on this new
hardware. At least they're reproducible.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>