hpcrun: Add timeout to Level Zero sync wait loop
Without a timeout, the Level Zero wait loop may hang the process if for some
reason the number of pending operations is miscounted.

Note that this is also evil but in a different way, while it "works" in most
cases it may lose data. So we really need to avoid this whole waiting thing
in the first place.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>