Add fallback for broken IGA 2.x packaging
Intel's packaging for the Intel Graphics Compiler (IGC) and Assembler
(IGA) is missing the unversioned symlink (lib*.so). This is very
unconventional, but work around the issue by looking for the current
versioned file (lib*.so.2) as a fallback.

The code tries to ensure the library it finds matches the headers, it
only searches the prefix where the headers are installed (assuming they
are installed under `{prefix}/include` like normal). It also searches
the "use this one dammit" environment variable IGA_ROOT.

Note that the version is hardcoded here, so if IGC 3.x releases
without this being fixed this code will likely do more harm than good.
Caveat emptor.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>