Don't force-enable any features
...Instead, use Meson's automatic detection to figure out what features
are available in each container environment.

This is intended to make configuration and expansion easier. Without the
added burden of updating all the native files, it is now very easy to
add a new feature to the build and, once the dependencies are available,
test it in CI.

The downside is that some tests can now be entirely skipped by the
pipeline, since a feature may accidentally never be enabled. Coverage
tracking is the (planned) better solution here, which will give a better
sense of what is and isn't tested anyway than dense config files.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>