Only look for GTPin if -Dgtpin is not disabled
Currently GTPin will be looked for even if the feature is disabled.
This causes the subproject to be loaded even if the libraries won't
actually be used at all. Add an extra conditional to ensure we only
download GTPin if we will actually use it.

This allows using Meson's --force-fallback-for option to allow/force
using the GTPin wrap even if `-Dgtpin=auto` (the default).

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>