hpcrun: Reimplement pthread_keys
pthread_key_create has known issues with multiple link namespaces, see
https://sourceware.org/bugzilla/show_bug.cgi?id=24776. The overlapping
keys manifest in a wide variety of ways, but sometimes result in crashes
especially when custom malloc/free implementations are in use.

This commit introduces a minimal pthread_key implementation that defers
the storage to a thread_local variable. While there are some minor
differences between this and "real" pthread_keys, it works well enough
to allow the private namespace to not cause more issues than it solves.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>