Fix Meson test dependencies on hpctesttool
Many of the tests depend on hpctesttool, but in Meson-land that is a
custom_idx object and due to a bug ([meson#14743]) `meson test` won't
automatically build it by default. This leads to confusion and
headaches when all the tests fail for no good reason. This commit
converts the `hpctesttool` into a custom_target which isn't affected by
the bug and thus "just works" like it should have from the start.

[meson#14743]: https://github.com/mesonbuild/meson/issues/14743

Fixes: https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/893
Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>