devcontainers: Add a remoteUser
This avoids user id issues when using rootful container runtimes such as
Docker. Without a `remoteUser` the container user is root on the host,
meaning that files created in the container cannot be edited by the
(host) user after the fact. With a `remoteUser` the container user has
the same user id as and maps to the host user.

Rootless runtimes such as Podman are not affected by this issue but are
not harmed by adding `remoteUser`. The primary change is now the
container user has the same user id as the host user.

With this change, the container user is now unprivileged inside the
container. This doesn't prevent a user from creating files with wrong
user ids on the host (i.e. privilege escalation), but it does block such
actions behind privileged tools such as `sudo` inside the container so
should be less likely to do by accident.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>