devcontainer: Set LANG in the container
`LANG` defines the locale to use by default when no locale is set.
Setting it here avoids errors where `LANG` is set to an unavailable
locale inside the devcontainer when run from VSCode.

Containers generally have most of their locales stripped for size, most
base containers also include code in `/etc/profile.d` that will confirm
the locale set in `LANG` (if set) is available and adjust it if needed.
The VSCode server will set `LANG` if it is not set already, however it
does so after the code in `/etc/profile.d` runs. The end result is that
VSCode terminals and tasks run with a `LANG` that is unavailable inside
the container.

This commit propagates `LANG` from the host environment, defaulting to
the only locale guaranteed to exist on Linux: `C.UTF-8`. This ensures
`LANG` is always set, which prevents the VSCode server from setting it
to an invalid value and thus avoids errors with e.g. Sphinx.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>