hpcrun: Add option to --disable-auditor by default
Some systems still have versions of Glibc that fail hard with LD_AUDIT.
In these scenarios it makes sense for --disable-auditor to be the
default, rather than --enable-auditor. This commit adds an option to
Meson setup to decide which should be the default at configuration time.

This replaces the previous logic that made --disable-auditor the default
on ARM systems. Now --enable-auditor is the default for all architectures
unless `-Dhpcrun_use_auditor_by_default=false` at Meson setup time.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>