lean/collections: Test rotations on duplicate insert
https://gitlab.com/hpctoolkit/hpctoolkit/-/merge_requests/1256 uncovered
a bug in this splay tree code: if an insert attempts to insert a key
already in the tree, parts of the tree may be lost due to a missing
update to the root pointer. The corresponding unit test did not cover
this case since it inserted duplicate keys consecutively, so no rotation
is invoked on the duplicate's insertion.

This commit reorders the duplicates to come after the "original"
insertions, ensuring at least one (and likely many) rotations during
duplicate insertions.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>