Remove ci/spack
The `spack` CI job(s) were intended to automatically detect when changes
to our Spack recipe are necessary, for instance because some dependency
was added or some build feature doesn't work as expected, etc.
Unfortunately, it has been broken for a long while due to factors not
entirely in our control. This commit removes the jobs entirely, with the
hope that a better alternative will eventually appear to take their place.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>