Use default CUDA architecture for test binaries
Previously, we would ask Meson for information on the available GPU
architectures to ensure the GPU in use would always be able to run the
tests. Meson's tables have not been updated for CUDA 13, so this
approach doesn't work for newer CUDA versions.

This commit removes this lookup and uses the default GPU architecture,
whatever that may be: 5.2 (Maxwell) for CUDA 11.x and 12.x, and
7.5 (Turing) for CUDA 13.x. Older Nvidia GPUs are either EOL or have
lost CUDA compatibility, so this change should not impact current users
(except making the build a little faster).

Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/901

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>