ci: Build for visible Nvidia GPUs
The default is to compile for an older but still usable compute
architecture. Unfortunately, this default seems to fail if the Nvidia
driver on the system is older than the CUDA toolkit running, with some
obtuse error related to incompatible PTX toolchains.

Why is this an issue when Nvidia explicitly says it isn't? Very unclear.
As a workaround, this commit configures the CI to use the "native" CUDA
architecture, which avoids the PTX entirely and thus always works.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>