Install Sphinx and dependencies with Pip
...as a fallback, if they aren't already available.

Sphinx is a complicated dependency: not only does it have depedencies of
its own, but we also use extensions which must be installed as Python
packages and have their own version relations with Sphinx core. This is
maddness that only a package manager can make sense of.

This commit punts the entire issue to Pip, Python's package manager. The
packages we need are written in `requirements.txt` format, we then use a
`pip install --dry-run` to confirm that they are installed. If not, we
install them inside a Python virtual environment.

The Pip wrapper logic respects Meson's `--wrap-mode` option, so there
(should be) no network access unless wraps are fully enabled.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>