Increase test timeouts for slow tests
Many tests are slow. There are many reasons why these tests are slow,
and in many cases it results from system differences that are difficult
to predict.

This commit increases (and often disables) the test timeouts for tests
that use functions known to be slow on some systems, such as POSIX
signals, multithreading, nvdisasm, and Python.

Timeouts are now always marked with a comment rationalizing the change
from the default timeout (30 seconds). This helps keep this very tribal
knowledge around for a bit longer.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>