Generate a CI pipeline performance report
Sometimes we need to assess the performance of the CI pipeline(s), but
gathering all the data needed for this can be an arduous process.
This commit introduces a new script that extracts this data
automatically, and generates a report that highlights key metrics of
interest for optimization.

Currently this script just does a cost breakdown for successful jobs,
but can be extended later to fit more use cases.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>