docs: Add EPUB and PDF links to the manual
With this change, the EPUB and PDF formats produced by Sphinx are
available from the manual itself, removing the need for an extra landing
page or links to provide these formats.

These links are by nature absolute, meaning they need to be adjusted
based on where the site is finally deployed. The default is to use the
install prefix, however this can be configured via a new Meson option
for deployment to GitLab Pages.

To make the build directory layout match the final site layout, the
Sphinx build has been significantly rewritten. Sphinx now builds
directly into the `doc/src/` build directory and shares this directory
across all 3 build outputs (and `latexmk`). This reduces the amount of
Meson/Sphinx shim code and removes repeated source parsing, making it
faster overall.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>