docs: Add a server for development
The Sphinx site can generate absolute links, this is most apparent in
the new EPUB and PDF links. The path we provide as the "base" for these
absolute links is the install prefix, which makes the install process
simple. Unfortunately, it also means parts of the page might break if
you attempt to view the page straight out of the build directory.

This commit adds a miniature HTTP server, based on Python's built-in
HTTP support, that maps the paths of requested files from the install
prefix to the build directory. The end result is a fully working site
suitable for development purposes.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>