docs: Add a version switcher
Our manual evolves over time, users will likely want to see a released
version of the manual matching their local install. This commit adds a
version switcher to make it easy for users to transfer to other versions
of the manual. The theme also adds a banner at the top when the version
is newer or older than the "preferred" latest release.

The version switcher is optional and does not appear unless the URL of
the version list JSON is provided to `meson setup`. CI/CD both passes
an appropriate URL and generates this JSON automatically.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>