docs: Compress installed artifacts
Web hosting services are often able to deliver the content in a
compressed format, provided the content is pre-compressed so the server
doesn't have to work for it. GitLab Pages is one such example. Extend
the HTML site installation process to automatically compress files.

Currently the only compression supported is gzip, which is both fast
and supported by standard Python.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>