docs: Refine branding and add favicon
SVGs are king when it comes to images in dynamic web environments,
replace our branding with its official SVG equivalent. The new image
also includes a white backdrop to prevent the wordmark borders from
disappearing when using the dark theme.

A favicon is also included, which appears in small at the top of the
browser window.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>