hpcrun: Strip R*PATHs from preload libraries
The `libhpcrun_preload*.so` libraries are designed to be loaded into the
application with `LD_PRELOAD`, they use symbol resolution to intercept
specific functions for measurement purposes. Because of this these
libraries are particularly sensitive to R*PATHs and especially RPATHs,
which alter and may conflict with the application's library searches.

Use `patchelf` on these particular libraries to remove any R*PATHs they
may have, either inadvertently from a clever compiler (wrapper) or as
defined by the user in `-D*_link_args` Meson options.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>