Fix Xed detection through CMake
On modern versions of Xed, the headers are stored under the `xed/`
subdirectory of the include path. While the package using our included
CMake script, it fails to extract the version from a nonexistent header
file. Meson considers a Xed without a proper version to not satisfy the
minimum version requirement and thus fail the dependency.

Fix `FindXed.cmake`'s version extraction to work when headers are stored
under a `xed/` prefix. This allows reusing late-model Xed from a
`CMAKE_PREFIX_PATH`.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>