ci: Verify plausible success before starting
There are a few jobs that run fairly fast and can easily halt the entire
pipeline, `pre-commit` and `clang-format` are the main examples. Instead
of starting up the pipeline only to cancel tens of jobs a minute later,
run these checks first and verify the basic requirements for the
pipeline are met before starting the larger jobs.

In the same vein, new `meson setup: [*]` jobs are added that check that
the build system is basically operational. These collapse and supercede
the more expensive `options: [*]` jobs, which had much the same goal but
also ran expensive compilations and tests with minimal benefit.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>