ci: Rename code quality to audit, drop clang-tidy-fix
`clang-tidy-fix` is the same as `clang-tidy`, but generates a patch full
of fix-its to apply locally to fix found errors. Most devs don't bother
and so this time is wasted, leave it out.

To generalize the pipeline stages a little more, the "code quality"
stage is renamed to "audit." Similar idea but more general, it can
eventually hold jobs for code coverage analysis among other quality
checks. For the most part these are expected to only document errors and
not enforce, so these jobs are naturally allowed to fail.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>