ci: Use largest available runners on GitLab
As a project in the GitLab OSS program, our credit usage is based
entirely on job duration and not the size of the runner we're running
on. Since most of our jobs have high internal parallelism, it's most
efficient to use the largest (and thus fastest) runners we have
available for these jobs.

Our group has access to GitLab Premium/Ultimate runners, but forks
don't. Previously we detected this by hardcoding the project ID, but we
can get the same effect cleaner using CI/CD variables. These variables
default to the appropriate value for forks, and then our group has
variables set to extend these to the largest available runners.

Using variables, the "hpctoolkit" and "gitlab" sites are nearly
identical, so we just merge them into one pipeline to keep things neat.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>