ci: Reduce spurious auto-cancels
Sometimes jobs fail and need to be retried. Canceling the pipeline
whenever one job fails is efficient if the user needs to make changes to
progress, but spurious failures require excessively spawning new
pipelines to work around the issue which is far from efficient.

Ideally, we would only auto-cancel jobs that are "safe" to cancel
(interruptible), but GitLab's auto-cancel configuration is limited here.
This commit uses the next best option: don't cancel the pipeline on job
failure.

It's hard to say whether this is more efficient than the prior version.
One nice feature is that failing pipelines give much more useful
information than before, which may reduce the number of `fixup!` commits
needed to pass CI.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>