ci: Reduce repeats for in test jobs
We tend to run all the tests in parallel anyway, repeating the tests
within each job burns a lot of time in comparison. Remove the repeats
for test jobs so they only run the tests once.

The exception here is aio, which still runs 7 repeats. Since aio is
expected to run all the tests, this ensures we still have reliable and
repeatable tests, without exposing flakiness across the pipeline.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>