hpcprof: Remove serialization in SparseDB
A major serialization lies in the SparseDB code, depending on the order
of parallel operations it is possible for the "prebuffer" to fill up
with unwritten metric data and end up seriously delaying the final
write() call.

It turns out this "prebuffer" is only needed to enforce a particular
section order in the final output, a property that is not actually
required the format nor any known readers. By relaxing this restriction,
we can avoid this significant slowdown and increase parallel efficiency.

As a side effect, PerThreadTemporaries (which contain per-profile metric
data) now have a much shorter, non-extendable lifetime in these cases.
This reduces the overall memory usage of hpcprof to the levels the
algorithm was originally designed to achieve.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>