hpctesttool: Refine diff, yaml commands for debugging
`hpctesttool diff LHS RHS` generates a diff from one database to another
or one database *file* to another, similar to the classic `diff` utility.
This is the new name for the `hpctesttool test db-compare` command,
although the recommended order for the arguments is reversed to match
normal `diff` semantics (i.e. canonical -> acquired).

`hpctesttool yaml DB OUTPUT` generates the YAML form for a database or
database file. This is the new name for the `hpctesttool test yaml`
command.

Both commands now accept individual database files, in addition to
complete databases.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>