Remove the api/ prefix from GTPin headers
GTPin is only available as a binary release and uses an extremely
non-standard layout with little to no documentation. Previously we
expected the `Include` subdirectory would be exposed on the include
path, so that our code would include `api/gtpin_api.h`. However Spack
exposes the headers themselves, i.e. without the `api/` prefix.

This commit changes our expectations to match what Spack gives us, i.e.
all the GTPin headers are exposed directly. So `cpp_args` should now be:

    -I/path/to/gtpin/Include/api -I/path/to/gtpin/Include/ged/intel64

This expectation matches most (but not all!) of the example code
included in the GTPin package.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>