devcontainer: Volume-mounted builddir/
Previously the devcontainer used a build directory embedded in a hidden
directory of the workspace bind-mount. While this works, it also exposes
binaries and other data that is fully container-specific to the host.
It is also non-portable, IDEs other than VSCode will not know to set
everything up in some subdirectory named by a hash.

The cleaner solution is a volume mount for the conventional `builddir/`.
This has the same persistance as the workspace bind-mount, but is
invisible to tools running inside the container and works with tools and
script that use the conventional `builddir/` for the build directory.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>