devcontainer: Portable ClangD support
ClangD by default cannot find the `compile_commands.json` generated by
Meson, because ClangD only looks in ancestors of the source files for
this file. Right now we provide configuration for the VSCode extension
to fix this, but other IDEs require configuring to make this work.

This commit adds a relative `compile_commands.json` symlink in the
workspace root which points into the conventional `builddir/` build
directory, if it doesn't exist already. With this symlink in place,
ClangD will always find the right compilation database even with no
configuration. This symlink is created as part of the initial container
startup and can be adjusted by the user afterwards if needed.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>