ci: Don't include debug info for compile jobs
Since the debug info is *by far* the largest size contributor to the
final build directory. We don't need debug info to run the tests, and
when we do we enable it explicitly.

This change has the added advantage of making the buildtype explicit
for all the jobs where it matters.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>