Control warnings for wrap fallbacks
We strive to develop quality software, which is why we regularly build
with `-Dwerror=true`. But while we might keep our space clean, not
everyone does, and a number of our wrap fallbacks generate a copious
number of compiler errors, and then fail to build when warnings are
treated as errors for our own testing.

This commit changes the default options passed to wrap fallbacks to
limit the effect their code quality has on our build. Wraps with known
issues now build with low or minimal warnings (`warning_level=0`), with
the worst offenders also having `werror=false`. These defaults only
apply on the first `meson setup` call and can be overridden with
`meson setup/configure -D<wrap>:<option>` like normal if needed.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>