subprojects: Add a Boost wrap
Historically, Boost hasn't been wrapped largely because the Boost build
system is a bit of a nightmarish mess. It turns out though that we +
Dyninst only need a small portion of this mess, and the portions we need
are more-or-less trivially wrapped.

This commit adds a wrap for Boost. This has the most dramatic effect on
RHEL 8 and OpenSUSE Leap 15, where the default Boost is 1.66 and too
old for Dyninst to use. With the wrap, Boost is now built automagically
instead of failing the build (or worse, causing headaches due to
incorrectly specified external Boost paths).

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>