meson: Remove hard dependency on host Python
We have a hard dependency on the "native" (build machine's) Python, but
don't actually need the host Python unless Python unwinding support is
enabled. This commit adjusts the Meson code to correctly use the native
Python across the build instead of the host Python, which may be
unavailable in cross builds (i.e. while cross-compiling).

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>