meson: Add Pip wrap
Pip is a hard dependency for `hpctesttool`, we need it to run the
install properly. Normally we can use Python's builtin `ensurepip`
module to acquire a Pip, but notably Debian/Ubuntu packages strip the
necessary wheel out by default. Add a wrap for Pip so the build will
still continue even without the necessary support.

Although we manually call subproject(), this respects the Meson builtin
options for fallbacks, i.e. it can be disabled with
`--wrap-mode=nofallback` and forced with `--force-fallback-for=pip`.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>