docs: Enhance search and fallback for Sphinx
In order to ensure the build will pass without a hitch, we check that
all the Sphinx extensions we want are installed at setup time.
Previously we used Pip to check this, but this requires that Sphinx is
installed in the "native" Python. This commit updates the check to use
Sphinx's own version-checking logic, using a "dummy" build to load up
enough of Sphinx to know whether it will work (or not).

This commit also enhances the Sphinx fallback, it now uses the same
simple technique as `hpctesttool`'s install. Since this is a fallback
it also respects Meson's builtin options for fallbacks, i.e. the
fallback can be disabled with `--wrap-mode=nofallback` or forced with
`--force-fallback-for=sphinx` as expected.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>