hpcrun: Record the exact instruction in the trace
Previously, we would record the (first PC of) enclosing function and not
the instruction offset where the sample actually occurred. This meant
any trace samples in an inlined function would instead be attributed to
the (non-inlined) caller function, leading to misleading traces that
blame the wrong function.

This commit instead records the exact instruction offset of the sample
in the trace, leaving it to `hpcviewer` to find the enclosing function.
(Which, as it happens, the Viewer seems to already do today.)

Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/908

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>