devcontainer: Drop CAP_SYS_ADMIN
For containers the seccomp profile is an important security feature
especially for rootful containers, and CAP_SYS_ADMIN permits far too
much. So there are security issues with leaving it open.

The more minimal alternative is CAP_PERFMON, this only enables access to
perf_event_open. Recent versions of Podman and Docker permit
perf_event_open with just CAP_PERFMON, so limit the devcontainer to just
that capability.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>