hpcstruct: Correct licensing for DTD
The DTD has it's own REUSE licensing declarations contained within.
These declarations should apply only to the DTD and not to the rest of
the Structfile, which should use a data-centric license (e.g. CDLA).
Express this in REUSE by marking the DTD as a "snippet", i.e. a fragment
with independent licensing declarations from the rest of the file.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>