hpcprof: Only combine non-zero data values
Most of hpcprof operates in "sparse" space, that is we only store the
measured data values that are not zero. The accumulators are a notable
exception to this rule, and as a result sometimes a zero can be the RHS
of an atomic combination operation. This is effectively a no-op for
addition, but for min() or max() this causes numerical errors.

Fix this by being more aggressive to ignore zeros, i.e. never combine a
value that was (originally) zero. In particular, this fixes cases where:

 1. Min would be reported too high (i.e. some values were ignored), and
 2. Count would count all threads instead of just the involved ones.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>