hpcprof: Generate statistics for invisible Metrics
ExtraStatistics currently only use the Sum statistic for Metrics, but
this isn't a principled limitation, it only exists because a similar
restriction exists in the `*.hpcrun` data format. Cleaning this up makes
the code generally easier to alter and update. And makes it possible to
use these other statistics in handwritten `metrics.yaml` formulas.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>