hpcstruct: fix hpcstruct --clean
The previous rules for removing hpcstruct and warnings
files didn't always work for GPU binaries because the
rules that removed $(GS) and $(GW) were sensitive to the
definition of GPUBIN_CFG = yes/no.

Replace the rules that precisely specify which hpcstruct
and warnings files to remove with rules that remove them
all using *.hpcstruct and *.warnings.

Signed-off-by: John Mellor-Crummey <johnmc@rice.edu>